api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
